refactor(ui): TMS-2587 rename view components
Due to how the app is turning out, it's too complicated for the route-level container components to be the only containers. Now allowing domain-specific container for when the route doesn't need to be involved.
