+++ frontend/src/components/group/GroupContainer.vue
