+++ frontend/src/components/tenant/TenantUserManagement.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { computed, ref } from 'vue'
 
-import GroupContainer from '@/components/container/GroupContainer.vue'
+import GroupContainer from '@/components/group/GroupContainer.vue'
 import ButtonPrimary from '@/components/ui/ButtonPrimary.vue'
 import ButtonSecondary from '@/components/ui/ButtonSecondary.vue'
 import FloatingActionButton from '@/components/ui/FloatingActionButton.vue'
