--- frontend/src/views/RolesView.vue
@@ -1,11 +1,0 @@
-<script setup lang="ts">
-import BaseSecureView from '@/views/BaseSecureView.vue'
-</script>
-
-<template>
-  <BaseSecureView>
-    <div>
-      <h1>My Roles</h1>
-    </div>
-  </BaseSecureView>
-</template>
